better primitive types

+++ lola/sparse.pxd
@@ -7,7 +7,7 @@ from libcpp.vector cimport vector as cppvector
 
 cdef class SparseCategorical:
 
-    cdef cppmap[int, float] _data
+    cdef cppmap[size_t, float] _data
     cdef float _base_value
     cdef size_t _support_size
 
@@ -17,13 +17,13 @@ cdef class SparseCategorical:
 
     cpdef size_t n_represented(self)
 
-    cpdef float get(self, int key)
+    cpdef float get(self, size_t key)
 
     cpdef void scale(self, float scalar)
 
     cpdef float sum(self)
 
-    cpdef float plus_equals(self, int key, float value)
+    cpdef float plus_equals(self, size_t key, float value)
 
     cpdef float normalise(self)
 
@@ -33,17 +33,17 @@ cdef class SparseCategorical:
 cdef class CPDTable:
 
     cdef:
-        cppvector[cppmap[int, float]] _cpds
+        cppvector[cppmap[size_t, float]] _cpds
         cppvector[float] _base_values
         size_t _support_size
 
-    cpdef float get(self, size_t x, int y)
+    cpdef float get(self, size_t x, size_t y)
 
     cpdef void scale(self, size_t x, float scalar)
 
     cpdef float sum(self, size_t x)
 
-    cpdef float plus_equals(self, size_t x, int y, float value)
+    cpdef float plus_equals(self, size_t x, size_t y, float value)
 
     cpdef void normalise(self)
